+++ src/Components/Editor/SidePanel/PanelTransform/Widgets/SnapVertices.tsx
